bee-movie-main/requirements.txt
Flask
Flask-Mail
google-auth-oauthlib
google-auth